✨ feat: product 권한 조회 로직 추가

+++ src/main/java/com/project/likelion14thbe/domain/product/service/command/ProductCommandServiceImpl.java
@@ -1,5 +1,7 @@
 package com.project.likelion14thbe.domain.product.service.command;
 
+import com.project.likelion14thbe.domain.order.execption.OrderErrorCode;
+import com.project.likelion14thbe.domain.order.execption.OrderException;
 import com.project.likelion14thbe.domain.product.converter.ProductConverter;
 import com.project.likelion14thbe.domain.product.dto.request.ProductReqDTO;
 import com.project.likelion14thbe.domain.product.dto.response.ProductResDTO;
@@ -29,19 +31,29 @@ public class ProductCommandServiceImpl implements ProductCommandService {
     }
 
     @Override
-    public void updateProduct(Long productId, ProductReqDTO.ProductChangeDTO dto){
+    public void updateProduct(Long productId, Long memberId, ProductReqDTO.ProductChangeDTO dto){
         // 상품 정보 조회
         Product product = productRepository.findByAndNotDeleted(productId)
                 .orElseThrow(() -> new ProductException(ProductErrorCode.PRODUCT_NOT_FOUND));
+
+        // 상품 정보 접근 권한 확인
+        if (!product.getMember().getId().equals(memberId)) {
+            throw new OrderException(OrderErrorCode.ORDER_FORBIDDEN);
+        }
 
         product.updateProduct(dto.productPrice(), dto.productQuantity(), dto.productName(), dto.description());
     }
 
     @Override
-    public void deleteProduct(Long productId){
+    public void deleteProduct(Long productId, Long memberId){
         //상품 정보 조회
         Product product = productRepository.findByAndNotDeleted(productId)
                 .orElseThrow(() -> new ProductException(ProductErrorCode.PRODUCT_NOT_FOUND));
+
+        // 리뷰 접근 권한 확인
+        if (!product.getMember().getId().equals(memberId)) {
+            throw new OrderException(OrderErrorCode.ORDER_FORBIDDEN);
+        }
 
         product.delete();
     }
